requirements.txt
discord-py>=2.5.2
email-validator>=2.2.0
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
python-dotenv>=1.1.0
aiohttp>=3.8.6
asyncio>=3.4.3
psutil>=5.9.5
requests>=2.31.0
discord.py
flask
gunicorn
discord.py
flask
gunicorn